+++ src/js/helpers.js
@@ -1,3 +1,9 @@
+$.extend($.expr[':'], {
+  'containsIN': function (elem, i, match, array) {
+    return (elem.textContent || elem.innerText || '').toLowerCase().indexOf((match[3] || '').toLowerCase()) >= 0
+  }
+})
+
 numeral.register('locale', 'pwyf', {
   delimiters: {
     thousands: ',',